GBE-679: extracted TechInfo cluster to separate file
